Only export Perl_alloccopstash under threads.
It might make more sense for makedef.pl to more carefully examine
embed.fnc for what's a symbol to be exported only under threads and
what's not, but for now, we appear to be maintaining a hard-coded
list of symbols not to export when not under threads, so let's add
Perl_alloccopstash to the list.  Failing to do so breaks the
build on platforms where the linker is literal-minded about what
symbols it's been told to find in the export list.