Don’t crash with ()=&CORE::srand
Ops that don’t allow arbitrarily long argument lists are flagged at compile-time with the number of children.

Coresubs have to call the same op with differing numbers of arguments,
so they push nulls on to the stack to make up the number of items the
pp function expects to pop.

Commit f914a6829 stopped popping the null of the stack.

Before:
$ ./perl -le 'print &CORE::srand'
1240765685

After:
$ ./perl -le 'print &CORE::srand'
Bus error

List assignment does the same thing, and makes it easier to
write a test.