Import CCHOME from the environment when building with GCC on Windows
Using the environment for such things is preferable to always having to
specify a CCHOME= command-line argument, and is akin to setting various
environment variables via vcvars32.bat for VC++ builds.

Also remove a blank line to make it more obvious that copying DLLs to the
t/ folder is part of the test-prep target. (This is one thing which
doesn't work if CCHOME has not been set and MinGW isn't installed in its
default location.)