Validate the module filename created for bareword require.
Validate the generated filename after translation of '::' to '/' (and
possible conversion from VMS to Unix format) to keep the code simpler.
Reject empty module names, module names starting with '/' or '.'  (ie
absolute paths, hidden files, and '..'), and module names containing NUL
bytes or '/.' (ie hidden files and '..').

Add a test for Perl_load_module(), and check that it now rejects module
names which fall foul of the above rules.