Split the guts of pp_require into S_require_version() and S_require_file().
This makes it clear that pp_require is effectively 2 disjoint functions,
and that all the local variables previously declared at the start of
pp_require actually belong exclusively to the file loading functionality.
It is now possible to declare name as const, allowing the compiler to verify
that its value is never changed.